+++ src/lib/components/table/MobileTitle.tsx
@@ -8,6 +8,7 @@ import { TableTitle } from "./TableTitle";
 interface MobileTitleProps {
   title: string;
   count: Option<number>;
+  showCount?: boolean;
   onViewMore?: () => void;
 }
 const cardProps = {
@@ -17,14 +18,22 @@ const cardProps = {
   borderRadius: "8px",
 };
 
-export const MobileTitle = ({ onViewMore, title, count }: MobileTitleProps) => (
-  <Flex
-    style={cardProps}
-    onClick={!count ? undefined : onViewMore}
-    opacity={!count ? 0.5 : 1}
-    bg="gray.900"
-  >
-    <TableTitle title={title} count={count} mb={0} />
-    <CustomIcon name="chevron-right" color="gray.600" />
-  </Flex>
-);
+export const MobileTitle = ({
+  title,
+  count,
+  showCount = true,
+  onViewMore,
+}: MobileTitleProps) => {
+  const isDisabled = count === 0;
+  return (
+    <Flex
+      style={cardProps}
+      onClick={isDisabled ? undefined : onViewMore}
+      opacity={isDisabled ? 0.5 : 1}
+      bg="gray.900"
+    >
+      <TableTitle title={title} count={count} showCount={showCount} mb={0} />
+      <CustomIcon name="chevron-right" color="gray.600" />
+    </Flex>
+  );
+};