+++ src/lib/pages/home/components/index.ts
@@ -1,2 +1,3 @@
+export * from "./CardInfo";
 export * from "./DevShortcut";
 export * from "./TopDecorations";