+++ src/lib/pages/interact/component/drawer/body/ModuleFunctionBody.tsx
@@ -4,6 +4,7 @@ import { useCallback, useMemo, useState } from "react";
 
 import { ModuleEmptyState, NoImageEmptyState } from "../../common";
 import type { ModuleSelectFunction } from "../types";
+import { AmpEvent, track } from "lib/amplitude";
 import InputWithIcon from "lib/components/InputWithIcon";
 import { CountBadge } from "lib/components/module/CountBadge";
 import { FunctionCard } from "lib/components/module/FunctionCard";
@@ -73,6 +74,9 @@ export const ModuleFunctionBody = ({
   const onFunctionSelect = useCallback(
     (fn: ExposedFunction) => {
       if (module) {
+        track(AmpEvent.USE_MODULE_SELECTION_FUNCTION, {
+          functionType: fn.is_view ? "view" : "Execute",
+        });
         handleModuleSelect(module, fn);
         closeModal();
       }
@@ -95,6 +99,7 @@ export const ModuleFunctionBody = ({
             {module.moduleName}
           </Heading>
           <InputWithIcon
+            action="Module Selection Drawer Function Search"
             iconPosition="start"
             value={keyword}
             onChange={(e) => setKeyword(e.target.value)}