+++ src/lib/services/accountService.ts
@@ -2,12 +2,24 @@ import type { UseQueryOptions, UseQueryResult } from "@tanstack/react-query";
 import { useQuery } from "@tanstack/react-query";
 import { useCallback } from "react";
 
-import { useCelatoneApp, CELATONE_QUERY_KEYS } from "lib/app-provider";
+import {
+  useCelatoneApp,
+  CELATONE_QUERY_KEYS,
+  useBaseApiRoute,
+  useWasmConfig,
+} from "lib/app-provider";
 import {
   getAccountIdByAddressQueryDocument,
   getAccountTypeByAddressQueryDocument,
 } from "lib/query";
 import type { AccountType, Addr, Nullable, Option } from "lib/types";
+
+import {
+  getAccountInfo,
+  type AccountInfo,
+  getAccountTableCounts,
+  type AccountTableCounts,
+} from "./account";
 
 export const useAccountId = (
   walletAddress: Option<Addr>
@@ -67,5 +79,28 @@ export const useAccountType = (
       retry: 1,
       refetchOnWindowFocus: false,
     }
+  );
+};
+
+export const useAccountInfo = (address: Addr): UseQueryResult<AccountInfo> => {
+  const endpoint = useBaseApiRoute("accounts");
+
+  return useQuery(
+    [CELATONE_QUERY_KEYS.BALANCES, endpoint, address],
+    async () => getAccountInfo(endpoint, address),
+    { enabled: !!address, retry: 1, refetchOnWindowFocus: false }
+  );
+};
+
+export const useAccountTableCounts = (
+  address: Addr
+): UseQueryResult<AccountTableCounts> => {
+  const endpoint = useBaseApiRoute("accounts");
+  const { enabled: isWasm } = useWasmConfig({ shouldRedirect: false });
+
+  return useQuery(
+    [CELATONE_QUERY_KEYS.TABLE_COUNTS, endpoint, address, isWasm],
+    async () => getAccountTableCounts(endpoint, address, isWasm),
+    { enabled: !!address, retry: 1, refetchOnWindowFocus: false }
   );
 };